feat: readOnly mode 추가
