--- stories/assets/components/StarRate.stories.tsx
@@ -1,24 +1,0 @@
-import type { Meta, StoryObj } from "@storybook/react";
-
-import StarRate from "@/components/StarRate";
-
-// More on how to set up stories at: https://storybook.js.org/docs/writing-stories#default-export
-const meta = {
-  title: "/components/StarRate",
-  component: StarRate,
-  parameters: {
-    // Optional parameter to center the component in the Canvas. More info: https://storybook.js.org/docs/configure/story-layout
-    layout: "centered",
-  },
-  // This component will have an automatically generated Autodocs entry: https://storybook.js.org/docs/writing-docs/autodocs
-  tags: ["autodocs"],
-  // More on argTypes: https://storybook.js.org/docs/api/argtypes
-} satisfies Meta<typeof StarRate>;
-
-export default meta;
-type Story = StoryObj<typeof meta>;
-
-// More on writing stories with args: https://storybook.js.org/docs/writing-stories/args
-export const Primary: Story = {};
-
-export const ReadOnly: Story = {};
